+++ AppSiba/app/src/main/java/com/example/co/appsiba/DB/model/IngreCate.java
@@ -2,22 +2,22 @@ package com.example.co.appsiba.db.model;
 
 public class IngreCate {
 
-    private Long id;
+    private int id;
     private String name;
 
     public IngreCate() {
     }
 
-    public IngreCate(Long id, String name) {
+    public IngreCate(int id, String name) {
         this.id = id;
         this.name = name;
     }
 
-    public Long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(int id) {
         this.id = id;
     }
